Added tower placement
Added a script called placeCube
makes a ghostcube that ignores rays that will mark the destination of where your tower will be placed, then when you click, it places the tower

Added a prefab and material called ghostcube. Ghostcube is a see through cube that follows your mouse.

Added an empty object called placeholder which contains the placeCube script

This way, you can click the place cube button, see where you're going to put it, then place a tower in the desired location